requirements.txt
python-docx
Jinja2
Pillow
requests
